Privacy-6: Verify offline-caching state by checking "about:config"
The old configuration menu for offline apps was removed, so we
need to migrate to the verification based on about:config.

Note on the sentence:

    または`offline.cache.capacity`の値が`1`である。

We need this, because turning off `offline.cache.enabled` has an
unexpexcted side-effect and we sometimes need to disable the offline
app cache by setting the capacity to an infinitesimal.