build-xlsx: enable to populate a form with text data
This makes build-xlsx a bit more advanced, and enables it to create
a "filled" spreadsheet using provided materials. Here is an example
invocation.

    $ ./build-xlsx esr68.txt esr60.txt

Each text file (esr*.txt) contains a set of specifications for the
Firefox version. The file format is same with the files in esr68/*.

Here is a minimum example:

    $ cat esr68.txt
    Admin-2: 初回起動時のアドオン導入警告の可否

    :1: 警告する（既定）

    lockPref("extensions.autoDisableScopes", 15);

This is a very experimental feature. Let's see how well this works.